+++ gen/c/std/crc32.c
@@ -2016,9 +2016,17 @@ uint32_t wuffs_crc32__ieee_hasher__update(wuffs_crc32__ieee_hasher* self,
   {
     wuffs_base__slice_u8 i_slice_p = a_x;
     uint8_t* v_p = i_slice_p.ptr;
-    uint8_t* i_end0_p = i_slice_p.ptr + (i_slice_p.len / 8) * 0;
+    uint8_t* i_end0_p = i_slice_p.ptr + (i_slice_p.len / 8) * 8;
     while (v_p < i_end0_p) {
       v_s ^= (wuffs_base__load_u32le(v_p + 0));
+      v_s = (wuffs_crc32__ieee_table[0][v_p[7]] ^
+             wuffs_crc32__ieee_table[1][v_p[6]] ^
+             wuffs_crc32__ieee_table[2][v_p[5]] ^
+             wuffs_crc32__ieee_table[3][v_p[4]] ^
+             wuffs_crc32__ieee_table[4][255 & (v_s >> 24)] ^
+             wuffs_crc32__ieee_table[5][255 & (v_s >> 16)] ^
+             wuffs_crc32__ieee_table[6][255 & (v_s >> 8)] ^
+             wuffs_crc32__ieee_table[7][255 & (v_s >> 0)]);
       v_p += 8;
     }
     uint8_t* i_end1_p = i_slice_p.ptr + i_slice_p.len;