+++ lang/check/bounds.go
@@ -855,6 +855,9 @@ func (q *checker) bcheckExprOther(n *a.Expr, depth uint32) (*big.Int, *big.Int, 
 		lengthExpr := (*a.Expr)(nil)
 		if lTyp := lhs.MType(); lTyp.IsArrayType() {
 			lengthExpr = lTyp.ArrayLength()
+		} else if lTyp.IsPtrType() {
+			// TODO: don't skip bounds checking p[i] for ptr-typed p.
+			break
 		} else {
 			lengthExpr = makeSliceLengthExpr(lhs)
 		}
@@ -874,6 +877,7 @@ func (q *checker) bcheckExprOther(n *a.Expr, depth uint32) (*big.Int, *big.Int, 
 			return nil, nil, nil
 		}
 
+		// TODO: slicing "ptr T" types?
 		lengthExpr := (*a.Expr)(nil)
 		if lTyp := lhs.MType(); lTyp.IsArrayType() {
 			lengthExpr = lTyp.ArrayLength()