+++ doc/changelog.md
@@ -19,6 +19,7 @@
 - Renamed `buf1`, `reader1`, etc to `io_buffer`, `io_reader`, etc.
 - Renamed `u32`, `io_reader`, etc to `base.u32`, `base.io_reader`, etc.
 - Added a `base.image_config` type.
+- Allow `p[i]` for pointer-typed `p`.
 - Tweaked how marks and limits work.
 - Supported animated (not just single frame) GIFs.
 - Marked the `std/gif` LZW decoder as private.